% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{cause_reg}
\alias{cause_reg}
\title{Print multiple regression results}
\usage{
cause_reg(cdata, covs = (1), digits = 2)
}
\arguments{
\item{cdata}{An object created with the cdata command}
}
\description{
Print multiple regression results
}
\examples{
cause_reg()
}

