% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{cdata}
\alias{cdata}
\title{Load data.}
\usage{
cdata(out, tx, cov = NULL, cov_type = c(1), family = "gaussian",
  design = "cohort")
}
\arguments{
\item{out}{A vector of outcomes}

\item{tx}{A numeric vector indicating whether a subject is "treatment" (1) or "control" (0)}

\item{cov}{A data frame of covariates}

\item{cov_type}{Vector indicating covariate types, 1=continuous, 2=ordinal, 3=categorical, defaults to continuous}

\item{family}{Distribution of outcome variable, defaults to "gaussian"}

\item{design}{Study design, defaults to "cohort'}
}
\description{
This function allows you to load your data into
}
\examples{
cdata()
}
\keyword{load}

