% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{des_covs2}
\alias{des_covs2}
\title{Explore covariates}
\usage{
des_covs2(cdata, covs = c(1), bin_no = 30, quints = 0,
  col_bar = "purple4", col_line = "#d3d3d3", theme = theme_bw())
}
\arguments{
\item{cdata}{An object created with the cdata command}

\item{covs}{A vector of covariate names to include, defaults to all}

\item{bin_no}{The number of histogram bins, defaults to 30}

\item{quints}{Adds quintiles to plots, 1 for control, 2 for treatment, defaults to none 0}

\item{col_bar}{The color of your histogram bars, defaults to "purple4"}

\item{col_line}{The color of your lines, defaults to "#d3d3d3"}

\item{theme}{Ggplot theme}
}
\description{
This function allows you to explore overlap and imbalance in outcomes by covariates
}
\examples{
des_covs2()
}

