% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{des_out}
\alias{des_out}
\title{Explore outcome distributions}
\usage{
des_out(cdata, bin_no = 30, col_bar = "purple4", col_line = "#d3d3d3",
  quints = TRUE, theme = t)
}
\arguments{
\item{cdata}{An object created with the cdata command}

\item{bin_no}{The number of histogram bins, defaults to 30}

\item{col_line}{The color of your lines, defaults to "#d3d3d3"}

\item{theme}{Ggplot theme}

\item{col}{The color of your histogram bars, defaults to "purple4"}
}
\description{
This function allows you to explore overlap in the distribution of outcomes by treatment and control
}
\examples{
des_out()
}

