% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{match.compare}
\alias{match.compare}
\title{Generate match output}
\usage{
match.compare(h)
}
\arguments{
\item{cdata}{An object created with the cdata command}
}
\description{
Generate match output
}
\examples{
match.compare()
}

