% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{reg_diag}
\alias{reg_diag}
\title{Examine regression diagnostics}
\usage{
reg_diag(cdata, covs = c(1), trim = c(NA))
}
\arguments{
\item{cdata}{An object created with the cdata command}

\item{covs}{A vector of covariate names to include, defaults to all}

\item{trim}{A logical vector used to trim the data set, defaults to keeping the entire dataset}
}
\description{
Examine regression diagnostics
}
\examples{
reg_diag()
}

